Factor out generic 'percentage' graph
